tests/wpt/tests/css/css-typed-om/the-stylepropertymap/properties/word-spacing.html

<!doctype html>
<meta charset="utf-8">
<title>'word-spacing' property</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-get">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-set">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#property-stle-value-normalization">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<script src="resources/testsuite.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

function assert_is_keyword_spacing(result, expected) {
  assert_style_value_equals(result, new CSSUnitValue(expected, 'px'));
}

runPropertyTests('word-spacing', [
  {
    syntax: 'normal',
    computed: (_, result) => assert_is_keyword_spacing(result, 0)
  },
  { syntax: '<length>' },
  { syntax: '<percentage>' },
]);

</script>
