third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/word-spacing-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'word-spacing' to the 'normal' keyword: normal
  assert_equals: expected "CSSKeywordValue" but got "CSSUnitValue"
Harness: the test ran to completion.
